Added documentation for SOAP functionality files
- Moved functionality testing code to the Functionality Tests directory.
