+++ SOAP/XML-parsing/XMLParser.h
@@ -1,6 +1,19 @@
-//
-// Created by Ross on 11/21/18.
-//
+//TODO: Implement this file and create a proper class out of the XMLParser functionality.
+//NOTE: All documentation is temporarily duplicated from XMLParser.cpp until
+//XMLParser.h is fully implemented.
+///Parses existing SOAP XML documents and generates new CMAC XML documents.
+/**
+ * This class has two primary functions: Reading and writing XML documents.
+ * The read functionality parses an existing SOAP XML document and stores
+ * it in memory for future use via a CAPMessage struct. The write functionality
+ * uses this struct to build a new CMAC XML document.
+ * Base solution modified from the following tutorial:
+ * https://akrzemi1.wordpress.com/2011/07/13/parsing-xml-with-boost/
+ *
+ * @file        XMLParser.h
+ * @authors     Ross Arcemont
+ * @date        November, 2018
+ */
 
 #ifndef CADG_XMLPARSER_H
 #define CADG_XMLPARSER_H
